requirements.txt
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.1
requests==2.32.3
playwright==1.46.0
web3
fastapi==0.112.0
uvicorn[standard]==0.30.3
pydantic==2.5.0
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
websockets==12.0
psutil==6.0.0
